fix: 修复 setSpeed 命令因 Float/Double 类型不匹配导致失效
MockServiceHelper.setSpeed 使用 putFloat 发送速度值，
但 RemoteCommandHandler 使用 getDouble 接收，
导致 Bundle 中类型不匹配，getDouble 始终返回默认值 0.0。
统一使用 putDouble 发送。

+++ app/src/main/java/moe/fuqiuluo/portal/service/MockServiceHelper.kt
@@ -201,7 +201,7 @@ object MockServiceHelper {
         }
         val rely = Bundle()
         rely.putString("command_id", "set_speed")
-        rely.putFloat("speed", speed)
+        rely.putDouble("speed", speed.toDouble())
         return locationManager.sendExtraCommand(PROVIDER_NAME, randomKey, rely)
     }
